change got_pack_parse_offset_delta tslen argument to size_t
The tslen argument is always in the (1..9) range so `int' is technically
fine (even int8_t would be!) but all the callers are passing a size_t,
so change the type accordingly.

ok stsp@

+++ lib/pack.c
@@ -963,7 +963,7 @@ parse_negative_offset(int64_t *offset, size_t *len, struct got_pack *pack,
 
 const struct got_error *
 got_pack_parse_offset_delta(off_t *base_offset, size_t *len,
-    struct got_pack *pack, off_t offset, int tslen)
+    struct got_pack *pack, off_t offset, size_t tslen)
 {
 	const struct got_error *err;
 	int64_t negoffset;
